+++ apps/dot-components-e2e/src/integration/Progress/Progress.spec.ts
@@ -1,12 +1,16 @@
 describe('dot-components: Progress component', () => {
-  before(() => cy.visit('/iframe.html?id=experimental-progress--default'));
+  before(() => cy.visit('/iframe.html?id=components-progress--default'));
 
   it('should have a dot- prefix', () => {
     cy.get('div').should('have.class', 'dot-progress');
   });
 
-  it('should render the component', () => {
+  it('should render a circular progress component', () => {
     cy.get('svg').should('have.class', 'MuiCircularProgress-svg');
+  });
+
+  it('should be rotating', () => {
+    cy.get('div').should('have.class', 'MuiCircularProgress-indeterminate');
   });
 
   describe('style decisions', () => {
@@ -14,7 +18,7 @@ describe('dot-components: Progress component', () => {
       cy.get('div.dot-progress').should(
         'have.css',
         'color',
-        'rgb(61, 108, 158)'
+        'rgb(100, 154, 61)'
       );
     });
   });