+++ libs/dot-components/src/lib/components/progress/Progress.stories.tsx
@@ -4,23 +4,17 @@ import { Story, Meta } from '@storybook/react/types-6-0';
 import { DotProgress, ProgressProps } from './Progress';
 
 export default {
-  title: 'Experimental/Progress',
+  title: 'Components/Progress',
   component: DotProgress,
   argTypes: {
-    color: { defaultValue: 'primary' },
     size: {
-      defaultValue: 40,
       control: { type: 'range', min: 15, max: 100, step: 1 },
     },
     thickness: {
-      defaultValue: 3.6,
       control: { type: 'range', min: 1, max: 10, step: 0.5 },
     },
     title: {
       defaultValue: 'Loading Data...',
-    },
-    variant: {
-      defaultValue: 'indeterminate',
     },
     value: {
       defaultValue: 20,